Not the right time to introduce this
Given that these students are not even beginner at this point, I feel like we need to stick to giving them just a few tools and ideas to learn to use. I do think this is interesting though and might make a good separate lecture near the end of the course?